requirements.txt
cupy-cuda12x
tqdm
-i https://pypi.nvidia.com
cudf-cu12
